refactor: split rust parser modules
