+++ crates/ox_content_parser/src/parser.rs
@@ -1,16 +1,29 @@
 //! Markdown parser implementation.
 
-use memchr::memchr;
-use ox_content_allocator::{Allocator, Vec};
-use ox_content_ast::{
-    AlignKind, BlockQuote, Document, Html, Image, Link, List, ListItem, Node, Paragraph, Span,
-    Table, TableCell, TableRow, Text,
-};
+use ox_content_allocator::Allocator;
+use ox_content_ast::{Document, Span};
 
-use crate::error::{ParseError, ParseResult};
+use crate::error::ParseResult;
 #[allow(unused_imports)]
 // The macro is no-op without the `profile` feature, which suppresses the use.
 use crate::profile_span;
+
+mod block;
+mod block_quote;
+mod cursor;
+mod fenced_code;
+mod html;
+mod inline;
+mod inline_helpers;
+mod inline_html;
+mod leaf;
+mod list;
+mod list_item;
+mod spans;
+mod table;
+
+#[cfg(test)]
+mod tests;
 
 /// Parser options.
 #[derive(Debug, Clone, Default)]
@@ -61,14 +74,6 @@ pub struct Parser<'a> {
     nesting_depth: usize,
 }
 
-struct ParsedListItem<'a> {
-    ordered: bool,
-    start: Option<u32>,
-    content: &'a str,
-    content_offset: usize,
-    checked: Option<bool>,
-}
-
 impl<'a> Parser<'a> {
     /// Creates a new parser with default options.
     #[must_use]
@@ -95,2095 +100,5 @@ impl<'a> Parser<'a> {
 
         let span = Span::new(0, self.source.len() as u32);
         Ok(Document { children, span })
-    }
-
-    /// Checks if we've reached the end of input.
-    fn is_at_end(&self) -> bool {
-        self.position >= self.source.len()
-    }
-
-    /// Returns the remaining source.
-    fn remaining(&self) -> &'a str {
-        &self.source[self.position..]
-    }
-
-    /// Peeks at the current character.
-    #[inline]
-    fn peek(&self) -> Option<char> {
-        // ASCII fast path: most parser hot loops only inspect ASCII bytes
-        // (`#`, `\``, `~`, `>`, digits, whitespace), so avoid the cost of
-        // constructing a `Chars` iterator and decoding UTF-8 when the
-        // current byte is plain ASCII.
-        let bytes = self.source.as_bytes();
-        let pos = self.position;
-        let &b = bytes.get(pos)?;
-        if b < 0x80 {
-            Some(b as char)
-        } else {
-            self.source[pos..].chars().next()
-        }
-    }
-
-    /// Advances by one character.
-    #[inline]
-    fn advance(&mut self) -> Option<char> {
-        let ch = self.peek()?;
-        self.position += ch.len_utf8();
-        Some(ch)
-    }
-
-    /// Skips whitespace characters.
-    fn skip_whitespace(&mut self) {
-        let bytes = self.source.as_bytes();
-        let mut pos = self.position;
-        while pos < bytes.len() && matches!(bytes[pos], b' ' | b'\t') {
-            pos += 1;
-        }
-        self.position = pos;
-    }
-
-    /// Skips blank lines.
-    fn skip_blank_lines(&mut self) {
-        let bytes = self.source.as_bytes();
-        let mut pos = self.position;
-        loop {
-            let line_start = pos;
-            // Skip spaces/tabs.
-            while pos < bytes.len() && matches!(bytes[pos], b' ' | b'\t') {
-                pos += 1;
-            }
-            if pos < bytes.len() && bytes[pos] == b'\n' {
-                pos += 1;
-            } else {
-                self.position = line_start;
-                return;
-            }
-        }
-    }
-
-    /// Parses a block element.
-    fn parse_block(&mut self) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_block");
-        self.skip_blank_lines();
-
-        if self.is_at_end() {
-            return Ok(None);
-        }
-
-        // Check nesting depth
-        if self.nesting_depth > self.options.max_nesting_depth {
-            return Err(ParseError::NestingTooDeep {
-                span: Span::new(self.position as u32, self.position as u32),
-                max_depth: self.options.max_nesting_depth,
-            });
-        }
-
-        let start = self.position;
-        let line = self.current_line();
-        let trimmed = line.trim_start();
-        let first = trimmed.as_bytes().first().copied();
-
-        // Try to parse different block types. Each dispatch arm hands the
-        // already-scanned `line` / `trimmed` slices to the matching
-        // `*_at` helper so the inner check doesn't re-run `memchr` +
-        // `trim_start` on the same line.
-        match first {
-            Some(b'#') if self.try_parse_heading_at(trimmed) => return self.parse_heading(start),
-            Some(b'-' | b'*') => {
-                if Self::try_parse_thematic_break_line(line) {
-                    return self.parse_thematic_break(start);
-                }
-                if Self::try_parse_list_line(trimmed) {
-                    return self.parse_list(start);
-                }
-            }
-            Some(b'_') if Self::try_parse_thematic_break_line(line) => {
-                return self.parse_thematic_break(start);
-            }
-            Some(b'>') => return self.parse_block_quote(start),
-            Some(b'`' | b'~') if Self::try_parse_fenced_code_at(line, trimmed) => {
-                return self.parse_fenced_code(start);
-            }
-            Some(b'<') if self.try_parse_html_block() => return self.parse_html_block(start),
-            Some(b'+' | b'0'..=b'9') if Self::try_parse_list_line(trimmed) => {
-                return self.parse_list(start);
-            }
-            _ => {}
-        }
-
-        if self.options.tables && memchr(b'|', line.as_bytes()).is_some() && self.try_parse_table()
-        {
-            return self.parse_table(start);
-        }
-
-        // Default: parse as paragraph
-        self.parse_paragraph(start)
-    }
-
-    /// Returns the current line (from `self.position` up to the next `\n`,
-    /// exclusive). Uses `memchr` for the newline scan — faster than
-    /// `remaining().lines().next()` because it operates on bytes instead
-    /// of UTF-8 code points.
-    fn current_line(&self) -> &'a str {
-        let bytes = self.source.as_bytes();
-        let end = memchr(b'\n', &bytes[self.position..])
-            .map_or(self.source.len(), |off| self.position + off);
-        &self.source[self.position..end]
-    }
-
-    fn line_starts_block(&self) -> bool {
-        let line = self.current_line();
-        let trimmed = line.trim_start();
-        let first = trimmed.as_bytes().first().copied();
-
-        // NB: dispatch arms must use the same checks as `parse_block`,
-        // which test at `self.position` (the un-trimmed offset). Using
-        // `is_atx_heading_prefix(trimmed.as_bytes())` here would diverge
-        // from `parse_block`'s `Some(b'#') if self.try_parse_heading_at()`
-        // check on indented inputs like `" # heading"`:
-        // `line_starts_block` would say "yes, a heading", `parse_block`
-        // would say "no" (since the byte at position 0 is a space) and
-        // fall through to `parse_paragraph`, which would immediately break
-        // with no content — `parse_block` then returns `Ok(None)` without
-        // advancing position, and the outer `parse()` loop spins forever
-        // on the same offset. We route every arm through the same
-        // line-cached helpers as `parse_block` so the two dispatchers stay
-        // in lock-step and we only `memchr` + `trim_start` the current
-        // line once per call.
-        let starts_block = match first {
-            Some(b'#') => self.try_parse_heading_at(trimmed),
-            Some(b'-' | b'*') => {
-                Self::try_parse_thematic_break_line(line) || Self::try_parse_list_line(trimmed)
-            }
-            Some(b'_') => Self::try_parse_thematic_break_line(line),
-            Some(b'>') => true,
-            Some(b'`' | b'~') => Self::try_parse_fenced_code_at(line, trimmed),
-            Some(b'<') => self.try_parse_html_block(),
-            Some(b'+' | b'0'..=b'9') => Self::try_parse_list_line(trimmed),
-            _ => false,
-        };
-
-        starts_block
-            || (self.options.tables
-                && memchr(b'|', line.as_bytes()).is_some()
-                && self.try_parse_table())
-    }
-
-    fn try_parse_html_block(&self) -> bool {
-        let line = self.remaining().lines().next().unwrap_or("");
-        Self::parse_html_block_tag_name(line).is_some() || line.trim_start().starts_with("<!--")
-    }
-
-    fn parse_html_block(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_html_block");
-        let line = self.remaining().lines().next().unwrap_or("");
-
-        if line.trim_start().starts_with("<!--") {
-            loop {
-                let consumed = self.consume_line();
-                if consumed.contains("-->") || self.is_at_end() {
-                    break;
-                }
-            }
-
-            let span = Span::new(start as u32, self.position as u32);
-            let value = &self.source[start..self.position];
-            return Ok(Some(Node::Html(Html { value, span })));
-        }
-
-        // `parse_html_block_tag_name` returns a borrowed slice — no allocation.
-        let Some(tag_name) = Self::parse_html_block_tag_name(line) else {
-            return Ok(None);
-        };
-
-        if Self::is_type1_html_block_tag(tag_name) {
-            // Type 1 blocks (`<pre>`, `<script>`, `<style>`, `<textarea>`)
-            // close on the first line containing `</tag`. The previous
-            // implementation built a `String` per parse + lowercased every
-            // line; instead we scan bytes directly with case-insensitive
-            // compare. For a single 25KB document with ~38 blocks this
-            // turns ~hundreds of allocations into zero.
-            let tag_bytes = tag_name.as_bytes();
-            loop {
-                let consumed = self.consume_line();
-                if ascii_contains_closing_tag(consumed, tag_bytes) || self.is_at_end() {
-                    break;
-                }
-            }
-        } else {
-            self.consume_line();
-            // Walk forward line-by-line until a blank line. Previously we
-            // peeked with `remaining().lines().next()` THEN consumed the
-            // line, which walked each line twice. Consume first and inspect
-            // the returned slice so we only scan once.
-            while !self.is_at_end() {
-                let line_start = self.position;
-                let consumed = self.consume_line();
-                if consumed.is_empty()
-                    || consumed.bytes().all(|b| b == b' ' || b == b'\t' || b == b'\r')
-                {
-                    // Roll back so the outer `parse_block` loop sees the
-                    // blank line and skips it via `skip_blank_lines`.
-                    self.position = line_start;
-                    break;
-                }
-            }
-        }
-
-        let span = Span::new(start as u32, self.position as u32);
-        let value = &self.source[start..self.position];
-        Ok(Some(Node::Html(Html { value, span })))
-    }
-
-    /// Returns the borrowed tag-name slice when `line` begins with one of
-    /// the recognized HTML block tags. The slice points back into `line`'s
-    /// underlying storage, so there's no allocation. Callers that need
-    /// case-insensitive comparison should use `eq_ignore_ascii_case` directly
-    /// — the source casing is preserved.
-    fn parse_html_block_tag_name(line: &str) -> Option<&str> {
-        let trimmed = line.trim_start();
-        let after_open = trimmed.strip_prefix('<')?;
-        let after_slash = after_open.strip_prefix('/').unwrap_or(after_open);
-        let mut tag_len = 0;
-
-        for byte in after_slash.as_bytes() {
-            if byte.is_ascii_alphanumeric() || *byte == b'-' {
-                tag_len += 1;
-            } else {
-                break;
-            }
-        }
-
-        if tag_len == 0 {
-            return None;
-        }
-
-        let tag_name = &after_slash[..tag_len];
-        let next = after_slash.as_bytes().get(tag_len).copied();
-
-        if let Some(byte) = next {
-            if !matches!(byte, b' ' | b'\t' | b'>' | b'/') {
-                return None;
-            }
-        }
-
-        if !Self::is_supported_html_block_tag(tag_name) {
-            return None;
-        }
-
-        Some(tag_name)
-    }
-
-    fn is_supported_html_block_tag(tag_name: &str) -> bool {
-        [
-            "article",
-            "aside",
-            "blockquote",
-            "details",
-            "dialog",
-            "div",
-            "figcaption",
-            "figure",
-            "footer",
-            "header",
-            "main",
-            "nav",
-            "ol",
-            "p",
-            "pre",
-            "script",
-            "section",
-            "style",
-            "summary",
-            "table",
-            "tbody",
-            "td",
-            "tfoot",
-            "th",
-            "thead",
-            "textarea",
-            "tr",
-            "ul",
-        ]
-        .iter()
-        .any(|candidate| tag_name.eq_ignore_ascii_case(candidate))
-    }
-
-    fn is_type1_html_block_tag(tag_name: &str) -> bool {
-        ["pre", "script", "style", "textarea"]
-            .iter()
-            .any(|candidate| tag_name.eq_ignore_ascii_case(candidate))
-    }
-
-    /// Parses a block quote.
-    fn parse_block_quote(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_block_quote");
-        self.nesting_depth += 1;
-
-        // Collect lines belonging to this block quote and strip the `>` prefix.
-        // Write straight into a bump-allocated `String` so we don't pay for
-        // `String::new` (system allocator) followed by `alloc_str` (copy to
-        // arena) on every block quote. Capacity is intentionally small —
-        // bumpalo will grow it if needed, and oversizing wastes arena
-        // bytes that can't be reclaimed until reset.
-        let bytes = self.source.as_bytes();
-        let mut inner = ox_content_allocator::String::with_capacity_in(128, self.allocator.bump());
-
-        loop {
-            if self.position >= bytes.len() {
-                break;
-            }
-
-            let line_start = self.position;
-            let mut ws_cursor = line_start;
-            while ws_cursor < bytes.len() && matches!(bytes[ws_cursor], b' ' | b'\t') {
-                ws_cursor += 1;
-            }
-
-            // Blank line ends the block quote
-            if ws_cursor >= bytes.len() || bytes[ws_cursor] == b'\n' {
-                break;
-            }
-
-            let line_end =
-                memchr(b'\n', &bytes[line_start..]).map_or(bytes.len(), |off| line_start + off);
-            let line = &self.source[line_start..line_end];
-            let trimmed_offset = ws_cursor - line_start;
-            let trimmed = &line[trimmed_offset..];
-
-            if let Some(after_gt) = trimmed.strip_prefix('>') {
-                // Strip the optional single space after `>`
-                let stripped = after_gt.strip_prefix(' ').unwrap_or(after_gt);
-                inner.push_str(stripped);
-                inner.push('\n');
-
-                // Advance past this line (and the trailing newline if any).
-                self.position = if line_end < bytes.len() { line_end + 1 } else { line_end };
-            } else {
-                // Line doesn't start with `>`, block quote ends
-                break;
-            }
-        }
-
-        // Recursively parse the inner content from the same arena — no copy.
-        let inner_str = inner.into_bump_str();
-        let sub_parser = Parser::with_options(self.allocator, inner_str, self.options.clone());
-        let sub_doc = sub_parser.parse()?;
-
-        self.nesting_depth -= 1;
-
-        let span = Span::new(start as u32, self.position as u32);
-        Ok(Some(Node::BlockQuote(BlockQuote { children: sub_doc.children, span })))
-    }
-
-    /// Checks if the current position starts a list.
-    fn try_parse_list(&self) -> bool {
-        Self::try_parse_list_line(self.current_line().trim_start())
-    }
-
-    /// Line-cached variant of [`Self::try_parse_list`]. Callers that
-    /// already produced the trimmed line via `parse_block` /
-    /// `line_starts_block` reuse that slice instead of re-scanning the
-    /// source for a newline and re-running `trim_start`.
-    fn try_parse_list_line(trimmed: &str) -> bool {
-        let trimmed = trimmed.as_bytes();
-
-        // Unordered list: starts with -, *, or + followed by space.
-        if trimmed.len() >= 2 && matches!(trimmed[0], b'-' | b'*' | b'+') && trimmed[1] == b' ' {
-            return true;
-        }
-
-        // Ordered list: digit(s) followed by `.` or `)` and a space.
-        let mut i = 0;
-        while i < trimmed.len() && trimmed[i].is_ascii_digit() {
-            i += 1;
-        }
-        i > 0
-            && i + 1 < trimmed.len()
-            && matches!(trimmed[i], b'.' | b')')
-            && trimmed[i + 1] == b' '
-    }
-
-    /// Calculates the indentation level (number of spaces) of the current line.
-    fn calc_indentation(&self, start: usize) -> usize {
-        let mut indent = 0;
-        let bytes = self.source.as_bytes();
-        for byte in bytes.iter().skip(start) {
-            match byte {
-                b' ' => indent += 1,
-                b'\t' => indent += 4, // Assume tab is 4 spaces
-                _ => break,
-            }
-        }
-        indent
-    }
-
-    fn parse_task_list_prefix(&self, content: &'a str) -> Option<(bool, usize)> {
-        if !self.options.task_lists || content.len() < 3 {
-            return None;
-        }
-
-        if (content.starts_with("[x]") || content.starts_with("[X]"))
-            && (content.len() == 3 || content.starts_with("[x] ") || content.starts_with("[X] "))
-        {
-            return Some((true, usize::from(content.len() > 3) + 3));
-        }
-
-        if content.starts_with("[ ]") && (content.len() == 3 || content.starts_with("[ ] ")) {
-            return Some((false, usize::from(content.len() > 3) + 3));
-        }
-
-        None
-    }
-
-    fn parse_list_item_line(&self, line_start: usize) -> Option<ParsedListItem<'a>> {
-        let remaining = &self.source[line_start..];
-        let line = remaining.lines().next().unwrap_or("");
-        self.parse_list_item_line_from_line(line_start, line)
-    }
-
-    fn parse_list_item_line_from_line(
-        &self,
-        line_start: usize,
-        line: &'a str,
-    ) -> Option<ParsedListItem<'a>> {
-        let trimmed = line.trim_start();
-        let trimmed_offset = line_start + (line.len() - trimmed.len());
-
-        if trimmed.starts_with("- ") || trimmed.starts_with("* ") || trimmed.starts_with("+ ") {
-            let mut content = &trimmed[2..];
-            let mut content_offset = trimmed_offset + 2;
-            let mut checked = None;
-
-            if let Some((done, consumed)) = self.parse_task_list_prefix(content) {
-                checked = Some(done);
-                content = &content[consumed..];
-                content_offset += consumed;
-            }
-
-            return Some(ParsedListItem {
-                ordered: false,
-                start: None,
-                content,
-                content_offset,
-                checked,
-            });
-        }
-
-        let bytes = trimmed.as_bytes();
-        let mut marker_end = 0;
-        while marker_end < bytes.len() && bytes[marker_end].is_ascii_digit() {
-            marker_end += 1;
-        }
-
-        if marker_end == 0 || marker_end + 1 >= bytes.len() {
-            return None;
-        }
-
-        let marker = bytes[marker_end];
-        if !matches!(marker, b'.' | b')') || bytes[marker_end + 1] != b' ' {
-            return None;
-        }
-
-        Some(ParsedListItem {
-            ordered: true,
-            start: trimmed[..marker_end].parse().ok(),
-            content: &trimmed[marker_end + 2..],
-            content_offset: trimmed_offset + marker_end + 2,
-            checked: None,
-        })
-    }
-
-    /// Returns the substring covering the line at `line_start`, excluding
-    /// the trailing `\n`. Equivalent to `self.source[line_start..].lines().next()`
-    /// but faster: `memchr` finds the newline in vectorized bytes rather
-    /// than walking UTF-8 code points one character at a time.
-    fn line_at(&self, line_start: usize) -> &'a str {
-        let bytes = self.source.as_bytes();
-        let end =
-            memchr(b'\n', &bytes[line_start..]).map_or(self.source.len(), |off| line_start + off);
-        &self.source[line_start..end]
-    }
-
-    fn next_line_start(&self, line_start: usize) -> usize {
-        let bytes = self.source.as_bytes();
-        match memchr(b'\n', &bytes[line_start..]) {
-            Some(off) => line_start + off + 1,
-            None => self.source.len(),
-        }
-    }
-
-    fn strip_indent_columns(line: &str, columns: usize) -> &str {
-        let mut consumed = 0;
-        let mut byte_index = 0;
-
-        for ch in line.chars() {
-            let width = if ch == ' ' {
-                1
-            } else if ch == '\t' {
-                4
-            } else {
-                break;
-            };
-
-            if consumed + width > columns {
-                break;
-            }
-
-            consumed += width;
-            byte_index += ch.len_utf8();
-        }
-
-        &line[byte_index..]
-    }
-
-    /// Parses a list (ordered or unordered).
-    fn parse_list(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_list");
-        let baseline_indent = self.calc_indentation(start);
-
-        // Determine list type from the first line (already verified by try_parse_list)
-        let first_line_start = self.position;
-        let Some(first_item) = self.parse_list_item_line(first_line_start) else {
-            return Ok(None);
-        };
-        let ordered = first_item.ordered;
-        let list_start = first_item.start;
-
-        let mut children: Vec<'a, ListItem<'a>> = self.allocator.new_vec();
-        let mut list_spread = false;
-
-        loop {
-            if self.is_at_end() {
-                break;
-            }
-
-            let line_start = self.position;
-            self.skip_whitespace();
-            if self.peek() == Some('\n') || self.is_at_end() {
-                self.position = line_start; // Backtrack to handle end of block
-                break;
-            }
-
-            // Check indentation
-            let current_indent = self.calc_indentation(line_start);
-
-            // If less indented, list ends
-            if current_indent < baseline_indent {
-                self.position = line_start;
-                break;
-            }
-
-            // If indented more, check if it's a nested list
-            if current_indent > baseline_indent {
-                // Peek to see if it's a list marker
-                self.position = line_start; // Reset position to check marker properly
-                if self.try_parse_list() {
-                    // Parse nested list
-                    if let Some(Node::List(nested_list)) = self.parse_list(line_start)? {
-                        // Add to the LAST item's children
-                        if let Some(last_item) = children.last_mut() {
-                            last_item.span = last_item.span.merge(nested_list.span);
-                            last_item.children.push(Node::List(nested_list));
-                        }
-                    }
-                } else {
-                    // Continuation content?
-                    // For now, we only support simple lists.
-                    // Just skip line to avoid infinite loop
-                    while let Some(ch) = self.peek() {
-                        self.advance();
-                        if ch == '\n' {
-                            break;
-                        }
-                    }
-                }
-                continue;
-            }
-
-            // Same indentation (or close enough? Standard is complex, we use strict >= baseline)
-            self.position = line_start; // Reset
-
-            // Check if it's a list item
-            let remaining = self.remaining();
-            let line = remaining.lines().next().unwrap_or("");
-            let Some(item) = self.parse_list_item_line_from_line(line_start, line) else {
-                break;
-            };
-            if item.ordered != ordered {
-                // Not a list item, break
-                break;
-            }
-
-            // Consume line
-            self.position += line.len();
-            let consumed_newline = self.peek() == Some('\n');
-            if consumed_newline {
-                self.advance();
-            }
-
-            let content_indent = item.content_offset.saturating_sub(line_start);
-            // Bump-allocate the per-item buffer so we don't go System →
-            // arena for every list item. Most items are a single short
-            // line — over-sizing the capacity wastes arena bytes that
-            // bumpalo can't reclaim until reset. Use the item header
-            // length as the floor and let push_str grow only when the
-            // item actually spans continuation lines.
-            let mut item_source = ox_content_allocator::String::with_capacity_in(
-                item.content.len() + 1,
-                self.allocator.bump(),
-            );
-            item_source.push_str(item.content);
-            if consumed_newline {
-                item_source.push('\n');
-            }
-            let mut item_end = self.position;
-            let mut item_spread = false;
-
-            loop {
-                if self.is_at_end() {
-                    break;
-                }
-
-                let continuation_start = self.position;
-                let continuation_line = self.line_at(continuation_start);
-                let continuation_next = self.next_line_start(continuation_start);
-
-                if continuation_line.trim().is_empty() {
-                    let mut lookahead = continuation_next;
-                    let mut blank_count = 1;
-                    while lookahead < self.source.len() {
-                        let line = self.line_at(lookahead);
-                        if !line.trim().is_empty() {
-                            break;
-                        }
-                        blank_count += 1;
-                        lookahead = self.next_line_start(lookahead);
-                    }
-
-                    if lookahead >= self.source.len() {
-                        break;
-                    }
-
-                    let next_indent = self.calc_indentation(lookahead);
-                    let next_item = self.parse_list_item_line(lookahead);
-                    if next_indent == baseline_indent
-                        && next_item.as_ref().is_some_and(|next| next.ordered == ordered)
-                    {
-                        self.position = lookahead;
-                        item_spread = true;
-                        list_spread = true;
-                        break;
-                    }
-
-                    if next_indent >= content_indent {
-                        for _ in 0..blank_count {
-                            item_source.push('\n');
-                        }
-                        self.position = lookahead;
-                        item_spread = true;
-                        list_spread = true;
-                        item_end = self.position;
-                        continue;
-                    }
-
-                    break;
-                }
-
-                let current_indent = self.calc_indentation(continuation_start);
-                if current_indent < baseline_indent {
-                    break;
-                }
-
-                if current_indent == baseline_indent {
-                    if self
-                        .parse_list_item_line(continuation_start)
-                        .is_some_and(|next| next.ordered == ordered)
-                    {
-                        break;
-                    }
-
-                    break;
-                }
-
-                let stripped = Self::strip_indent_columns(continuation_line, content_indent);
-                item_source.push_str(stripped);
-                item_source.push('\n');
-                self.position = continuation_next;
-                item_end = self.position;
-            }
-
-            let item_source = item_source.into_bump_str();
-            let sub_parser =
-                Parser::with_options(self.allocator, item_source, self.options.clone());
-            let sub_doc = sub_parser.parse()?;
-            let mut item_children = sub_doc.children;
-            for child in &mut item_children {
-                Self::offset_node_spans(child, item.content_offset as u32);
-            }
-
-            let list_item = ListItem {
-                checked: item.checked,
-                spread: item_spread,
-                children: item_children,
-                span: Span::new(line_start as u32, item_end as u32),
-            };
-            children.push(list_item);
-        }
-
-        let span = Span::new(start as u32, self.position as u32);
-        Ok(Some(Node::List(List {
-            ordered,
-            start: list_start,
-            spread: list_spread,
-            children,
-            span,
-        })))
-    }
-
-    fn offset_span(span: &mut Span, offset: u32) {
-        span.start += offset;
-        span.end += offset;
-    }
-
-    fn offset_node_spans(node: &mut Node<'a>, offset: u32) {
-        match node {
-            Node::Paragraph(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::Heading(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::ThematicBreak(node) => Self::offset_span(&mut node.span, offset),
-            Node::BlockQuote(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::List(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_list_item_spans(child, offset);
-                }
-            }
-            Node::ListItem(node) => Self::offset_list_item_spans(node, offset),
-            Node::CodeBlock(node) => Self::offset_span(&mut node.span, offset),
-            Node::Html(node) => Self::offset_span(&mut node.span, offset),
-            Node::Table(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for row in &mut node.children {
-                    Self::offset_span(&mut row.span, offset);
-                    for cell in &mut row.children {
-                        Self::offset_span(&mut cell.span, offset);
-                        for child in &mut cell.children {
-                            Self::offset_node_spans(child, offset);
-                        }
-                    }
-                }
-            }
-            Node::Text(node) => Self::offset_span(&mut node.span, offset),
-            Node::Emphasis(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::Strong(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::InlineCode(node) => Self::offset_span(&mut node.span, offset),
-            Node::Break(node) => Self::offset_span(&mut node.span, offset),
-            Node::Link(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::Image(node) => Self::offset_span(&mut node.span, offset),
-            Node::Delete(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-            Node::FootnoteReference(node) => Self::offset_span(&mut node.span, offset),
-            Node::Definition(node) => Self::offset_span(&mut node.span, offset),
-            Node::FootnoteDefinition(node) => {
-                Self::offset_span(&mut node.span, offset);
-                for child in &mut node.children {
-                    Self::offset_node_spans(child, offset);
-                }
-            }
-        }
-    }
-
-    fn offset_list_item_spans(list_item: &mut ListItem<'a>, offset: u32) {
-        Self::offset_span(&mut list_item.span, offset);
-        for child in &mut list_item.children {
-            Self::offset_node_spans(child, offset);
-        }
-    }
-
-    /// Cheap line-cached ATX heading check. Both `parse_block` and
-    /// `line_starts_block` already produce the trimmed line, so we reuse
-    /// that slice instead of re-scanning the source. ATX rejects any
-    /// leading indentation, so the heading test only fires when the line
-    /// at `self.position` does NOT start with a space or tab.
-    fn try_parse_heading_at(&self, trimmed: &str) -> bool {
-        let bytes = self.source.as_bytes();
-        let pos = self.position;
-        if pos < bytes.len() && matches!(bytes[pos], b' ' | b'\t') {
-            return false;
-        }
-        is_atx_heading_prefix(trimmed.as_bytes())
-    }
-
-    fn try_parse_thematic_break_line(line: &str) -> bool {
-        let bytes = line.trim().as_bytes();
-        if bytes.len() < 3 {
-            return false;
-        }
-        let first = bytes[0];
-        if !matches!(first, b'-' | b'*' | b'_') {
-            return false;
-        }
-        let mut count = 0u32;
-        for &b in bytes {
-            if b == first {
-                count += 1;
-            } else if b != b' ' && b != b'\t' {
-                return false;
-            }
-        }
-        count >= 3
-    }
-
-    fn try_parse_fenced_code_at(line: &str, trimmed: &str) -> bool {
-        if Self::indentation_columns(line) > 3 {
-            return false;
-        }
-
-        let trimmed = trimmed.as_bytes();
-        trimmed.len() >= 3
-            && ((trimmed[0] == b'`' && trimmed[1] == b'`' && trimmed[2] == b'`')
-                || (trimmed[0] == b'~' && trimmed[1] == b'~' && trimmed[2] == b'~'))
-    }
-
-    fn indentation_columns(line: &str) -> usize {
-        let mut indent = 0;
-        for &b in line.as_bytes() {
-            match b {
-                b' ' => indent += 1,
-                b'\t' => indent += 4,
-                _ => break,
-            }
-        }
-        indent
-    }
-
-    /// Checks if the current position starts a table.
-    fn try_parse_table(&self) -> bool {
-        // Avoid the `Vec` allocation: peek the first two lines directly via
-        // `memchr` and inspect them in place.
-        let bytes = self.source.as_bytes();
-        let p0 = self.position;
-        let nl0 = match memchr(b'\n', &bytes[p0..]) {
-            Some(off) => p0 + off,
-            None => return false,
-        };
-        let p1 = nl0 + 1;
-        if p1 >= bytes.len() {
-            return false;
-        }
-        let nl1 = memchr(b'\n', &bytes[p1..]).map_or(bytes.len(), |off| p1 + off);
-
-        let first_line = self.source[p0..nl0].trim();
-        if memchr(b'|', first_line.as_bytes()).is_none() {
-            return false;
-        }
-
-        // Second line must be the delimiter row (contains | and -)
-        let second_line = self.source[p1..nl1].trim();
-        if memchr(b'|', second_line.as_bytes()).is_none()
-            || memchr(b'-', second_line.as_bytes()).is_none()
-        {
-            return false;
-        }
-
-        // Check delimiter row pattern: |---|---|
-        let is_delimiter = second_line.split('|').filter(|s| !s.is_empty()).all(|cell| {
-            let trimmed = cell.trim();
-            if trimmed.is_empty() {
-                return true;
-            }
-            // Allow :---:, :---, ---:, ---
-            trimmed.chars().all(|c| c == '-' || c == ':')
-        });
-
-        is_delimiter
-    }
-
-    /// Parses a heading.
-    fn parse_heading(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_heading");
-        let mut depth = 0u8;
-        while self.peek() == Some('#') {
-            depth += 1;
-            self.advance();
-        }
-
-        self.skip_whitespace();
-
-        let content_start = self.position;
-        let mut content_end = content_start;
-
-        // Read until end of line
-        while let Some(ch) = self.peek() {
-            if ch == '\n' {
-                break;
-            }
-            self.advance();
-            content_end = self.position;
-        }
-
-        // Skip trailing hashes and whitespace
-        let content = self.source[content_start..content_end].trim_end();
-        let content = content.trim_end_matches('#').trim_end();
-
-        // Consume newline
-        if self.peek() == Some('\n') {
-            self.advance();
-        }
-
-        let span = Span::new(start as u32, self.position as u32);
-
-        // Parse inline content
-        let children = if !content.is_empty() {
-            self.parse_inline(content, content_start)?
-        } else {
-            self.allocator.new_vec()
-        };
-
-        Ok(Some(Node::Heading(ox_content_ast::Heading { depth, children, span })))
-    }
-
-    /// Parses a thematic break.
-    fn parse_thematic_break(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        // Skip to end of line
-        while let Some(ch) = self.peek() {
-            self.advance();
-            if ch == '\n' {
-                break;
-            }
-        }
-
-        let span = Span::new(start as u32, self.position as u32);
-        Ok(Some(Node::ThematicBreak(ox_content_ast::ThematicBreak { span })))
-    }
-
-    /// Scans for a closing fence with at least `fence_len` repetitions of
-    /// `fence_char`, allowing up to 3 columns of leading indent on the
-    /// candidate line. Returns `(body_end, fence_line_end)` where
-    /// `body_end` is the offset to slice the code body and
-    /// `fence_line_end` is the offset *after* the closing fence's newline
-    /// (or EOF). When no closing fence exists, returns `(eof, eof)` —
-    /// matching the legacy "consume to end of input" behavior.
-    fn find_fenced_close(
-        &self,
-        fence_char: char,
-        fence_len: usize,
-        body_start: usize,
-    ) -> (usize, usize) {
-        let bytes = self.source.as_bytes();
-        let fence_byte = fence_char as u8;
-        let mut line_start = body_start;
-
-        while line_start < bytes.len() {
-            // Skip up to 3 leading spaces.
-            let mut cursor = line_start;
-            let max_indent_end = (line_start + 3).min(bytes.len());
-            while cursor < max_indent_end && bytes[cursor] == b' ' {
-                cursor += 1;
-            }
-
-            // Count the run of `fence_char`.
-            let fence_start = cursor;
-            while cursor < bytes.len() && bytes[cursor] == fence_byte {
-                cursor += 1;
-            }
-            let count = cursor - fence_start;
-
-            if count >= fence_len {
-                // Found the closing fence. Body ends at `line_start`;
-                // fence line ends at the next `\n` (inclusive) or EOF.
-                let after_fence = match memchr(b'\n', &bytes[cursor..]) {
-                    Some(off) => cursor + off + 1,
-                    None => bytes.len(),
-                };
-                return (line_start, after_fence);
-            }
-
-            // Not a closing fence — move to the next line.
-            line_start = match memchr(b'\n', &bytes[line_start..]) {
-                Some(off) => line_start + off + 1,
-                None => bytes.len(),
-            };
-        }
-
-        // No closing fence; consume everything as body.
-        (bytes.len(), bytes.len())
-    }
-
-    /// Parses a fenced code block.
-    fn parse_fenced_code(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_fenced_code");
-        let opening_indent = self.calc_indentation(start).min(3);
-        for _ in 0..opening_indent {
-            if self.peek() == Some(' ') {
-                self.advance();
-            }
-        }
-
-        let Some(fence_char) = self.peek() else {
-            return Err(ParseError::UnexpectedEof { span: Span::new(start as u32, start as u32) });
-        };
-        let mut fence_len = 0;
-
-        while self.peek() == Some(fence_char) {
-            fence_len += 1;
-            self.advance();
-        }
-
-        // Parse info string (language)
-        self.skip_whitespace();
-        let info_start = self.position;
-        while let Some(ch) = self.peek() {
-            if ch == '\n' {
-                break;
-            }
-            self.advance();
-        }
-        let info = self.source[info_start..self.position].trim();
-        let (lang, meta) = if info.is_empty() {
-            (None, None)
-        } else if let Some(space_idx) = info.find(' ') {
-            (Some(&info[..space_idx]), Some(&info[space_idx + 1..]))
-        } else {
-            (Some(info), None)
-        };
-
-        // Skip newline after info string
-        if self.peek() == Some('\n') {
-            self.advance();
-        }
-
-        // Fast path: when the opening fence has no indentation, the body
-        // lines need no indent stripping — we can find the closing fence
-        // by scanning the source and emit a zero-copy `&str` slice. This
-        // is the overwhelmingly common case (almost no code block in the
-        // wild is indented), and it removes both the per-line copy into a
-        // growing `String` *and* the trailing `alloc_str` (which used to
-        // double-allocate every code block's body).
-        let body_start = self.position;
-        let span;
-        let value: &'a str = if opening_indent == 0 {
-            let (body_end, fence_line_end) =
-                self.find_fenced_close(fence_char, fence_len, body_start);
-            self.position = fence_line_end;
-            span = Span::new(start as u32, self.position as u32);
-            &self.source[body_start..body_end]
-        } else {
-            // Indented opening fence: lines may need leading-space stripping,
-            // so we have to materialize a new string. Write directly into a
-            // bump-allocated string so we don't pay for `String` (system
-            // allocator) → `alloc_str` (arena copy).
-            let remaining_estimate = self.source.len().saturating_sub(self.position);
-            let mut value = ox_content_allocator::String::with_capacity_in(
-                remaining_estimate.min(8 * 1024),
-                self.allocator.bump(),
-            );
-
-            loop {
-                if self.is_at_end() {
-                    break;
-                }
-
-                let line_start = self.position;
-                let line = self.line_at(line_start);
-                let line_indent = Self::indentation_columns(line);
-
-                if line_indent <= 3 {
-                    self.position = line_start;
-                    let indent_to_skip = self.calc_indentation(line_start).min(3);
-                    for _ in 0..indent_to_skip {
-                        if self.peek() == Some(' ') {
-                            self.advance();
-                        }
-                    }
-
-                    // Check for closing fence
-                    let mut closing_fence_len = 0;
-                    while self.peek() == Some(fence_char) {
-                        closing_fence_len += 1;
-                        self.advance();
-                    }
-
-                    if closing_fence_len >= fence_len {
-                        // Skip rest of line
-                        while let Some(ch) = self.peek() {
-                            if ch == '\n' {
-                                self.advance();
-                                break;
-                            }
-                            self.advance();
-                        }
-                        break;
-                    }
-                }
-
-                // Not a closing fence, reset and consume line
-                self.position = line_start;
-                let next_line = self.next_line_start(line_start);
-                let stripped = Self::strip_indent_columns(line, opening_indent);
-                value.push_str(stripped);
-                if self.source.as_bytes().get(line_start + line.len()) == Some(&b'\n') {
-                    value.push('\n');
-                }
-                self.position = next_line;
-            }
-
-            span = Span::new(start as u32, self.position as u32);
-            value.into_bump_str()
-        };
-
-        Ok(Some(Node::CodeBlock(ox_content_ast::CodeBlock { lang, meta, value, span })))
-    }
-
-    /// Parses a table.
-    fn parse_table(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_table");
-        let mut align: Vec<'a, AlignKind> = self.allocator.new_vec();
-
-        // Parse header row
-        let header_line = self.consume_line();
-
-        // Parse delimiter row to get alignment
-        let delimiter_line = self.consume_line();
-        for cell in delimiter_line.split('|').filter(|s| !s.trim().is_empty()) {
-            let cell = cell.trim();
-            let starts_colon = cell.starts_with(':');
-            let ends_colon = cell.ends_with(':');
-            let alignment = match (starts_colon, ends_colon) {
-                (true, true) => AlignKind::Center,
-                (true, false) => AlignKind::Left,
-                (false, true) => AlignKind::Right,
-                (false, false) => AlignKind::None,
-            };
-            align.push(alignment);
-        }
-
-        // Build the table AST directly instead of first collecting row
-        // slices into short-lived heap Vecs.
-        let mut children: Vec<'a, TableRow<'a>> = self.allocator.new_vec();
-        children.push(self.parse_table_row(header_line)?);
-
-        // Parse body rows
-        loop {
-            if self.is_at_end() {
-                break;
-            }
-
-            let line_start = self.position;
-            self.skip_whitespace();
-
-            // Check for blank line or non-table line
-            if self.peek() == Some('\n') || self.is_at_end() {
-                self.position = line_start;
-                break;
-            }
-
-            // Check if line contains | (table continuation)
-            let remaining = self.remaining();
-            let line = remaining.lines().next().unwrap_or("");
-            if !line.contains('|') {
-                self.position = line_start;
-                break;
-            }
-
-            self.position = line_start;
-            let row_line = self.consume_line();
-            children.push(self.parse_table_row(row_line)?);
-        }
-
-        let span = Span::new(start as u32, self.position as u32);
-        Ok(Some(Node::Table(Table { align, children, span })))
-    }
-
-    /// Consumes a line and returns it.
-    ///
-    /// Uses `memchr` to find the terminator in vectorized bytes instead
-    /// of scanning per-byte. On a document like `docs/content/api/types.md`
-    /// (~25 KB, 38 HTML blocks) this is on the parser's hottest path.
-    fn consume_line(&mut self) -> &'a str {
-        let start = self.position;
-        let bytes = self.source.as_bytes();
-        if let Some(off) = memchr(b'\n', &bytes[start..]) {
-            let line_end = start + off;
-            self.position = line_end + 1;
-            &self.source[start..line_end]
-        } else {
-            self.position = self.source.len();
-            &self.source[start..]
-        }
-    }
-
-    /// Parses a table row into arena-backed AST cells without temporary heap
-    /// collection.
-    fn parse_table_row(&self, line: &'a str) -> ParseResult<TableRow<'a>> {
-        let mut cells: Vec<'a, TableCell<'a>> = self.allocator.new_vec();
-        for cell_content in Self::table_row_cells(line) {
-            let cell_children = self.parse_inline(cell_content, 0)?;
-            cells.push(TableCell { children: cell_children, span: Span::new(0, 0) });
-        }
-        Ok(TableRow { children: cells, span: Span::new(0, 0) })
-    }
-
-    /// Iterates table row cells from a line.
-    fn table_row_cells(line: &'a str) -> impl Iterator<Item = &'a str> {
-        let trimmed = line.trim();
-        let trimmed = trimmed.strip_prefix('|').unwrap_or(trimmed);
-        let trimmed = trimmed.strip_suffix('|').unwrap_or(trimmed);
-        trimmed.split('|').map(str::trim)
-    }
-
-    /// Parses a paragraph.
-    fn parse_paragraph(&mut self, start: usize) -> ParseResult<Option<Node<'a>>> {
-        profile_span!("parser::parse_paragraph");
-        let mut content_end = start;
-        let bytes = self.source.as_bytes();
-
-        loop {
-            if self.is_at_end() {
-                break;
-            }
-
-            // Check for blank line (paragraph end): scan whitespace and
-            // peek the next byte. Cheaper than the prior
-            // `skip_whitespace` + `peek` + reset dance.
-            let line_start = self.position;
-            let mut cursor = line_start;
-            while cursor < bytes.len() && matches!(bytes[cursor], b' ' | b'\t') {
-                cursor += 1;
-            }
-            if cursor >= bytes.len() || bytes[cursor] == b'\n' {
-                break;
-            }
-
-            // Check for block-level element that would end paragraph.
-            if self.line_starts_block() {
-                break;
-            }
-
-            // Consume one line via memchr.
-            content_end = if let Some(off) = memchr(b'\n', &bytes[line_start..]) {
-                line_start + off + 1
-            } else {
-                self.source.len()
-            };
-            self.position = content_end;
-        }
-
-        let content = self.source[start..content_end].trim();
-        if content.is_empty() {
-            return Ok(None);
-        }
-
-        let span = Span::new(start as u32, content_end as u32);
-
-        // Parse inline content
-        let children = self.parse_inline(content, start)?;
-
-        Ok(Some(Node::Paragraph(Paragraph { children, span })))
-    }
-
-    /// Parses inline content.
-    fn parse_inline(&self, content: &'a str, offset: usize) -> ParseResult<Vec<'a, Node<'a>>> {
-        profile_span!("parser::parse_inline");
-        let mut children = self.allocator.new_vec();
-        let mut pos = 0;
-        let bytes = content.as_bytes();
-
-        while pos < content.len() {
-            let start = pos;
-
-            // Fast-skip to the next special character using an unrolled
-            // byte test. The hot path in real-world Markdown is runs of
-            // plain text — this lets the optimizer vectorize the scan.
-            pos = next_inline_special(bytes, pos);
-
-            // Emit text before special character
-            if pos > start {
-                let text_content = &content[start..pos];
-                let text = Text {
-                    value: text_content,
-                    span: Span::new((offset + start) as u32, (offset + pos) as u32),
-                };
-                children.push(Node::Text(text));
-            }
-
-            if pos >= content.len() {
-                break;
-            }
-
-            // Handle special characters
-            let ch = bytes[pos];
-            match ch {
-                b'\\' if pos + 1 < content.len() && bytes[pos + 1] == b'\n' => {
-                    let break_node = ox_content_ast::Break {
-                        span: Span::new((offset + pos) as u32, (offset + pos + 2) as u32),
-                    };
-                    children.push(Node::Break(break_node));
-                    pos += 2;
-                }
-                b'<' => {
-                    if let Some((html, end)) = Self::parse_inline_html(content, pos, offset) {
-                        children.push(Node::Html(html));
-                        pos = end;
-                    } else {
-                        let text = Text {
-                            value: "<",
-                            span: Span::new((offset + pos) as u32, (offset + pos + 1) as u32),
-                        };
-                        children.push(Node::Text(text));
-                        pos += 1;
-                    }
-                }
-                b'\\' if pos + 1 < content.len() => {
-                    // Escape sequence
-                    pos += 1;
-                    let escaped = &content[pos..pos + 1];
-                    let text = Text {
-                        value: escaped,
-                        span: Span::new((offset + pos - 1) as u32, (offset + pos + 1) as u32),
-                    };
-                    children.push(Node::Text(text));
-                    pos += 1;
-                }
-                b'~' if self.options.strikethrough
-                    && pos + 1 < content.len()
-                    && bytes[pos + 1] == b'~' =>
-                {
-                    let inner_start = pos + 2;
-                    let mut inner_end = inner_start;
-                    let mut found = false;
-
-                    while inner_end + 1 < content.len() {
-                        if bytes[inner_end] == b'~' && bytes[inner_end + 1] == b'~' {
-                            found = true;
-                            break;
-                        }
-                        inner_end += 1;
-                    }
-
-                    if found {
-                        let inner_content = &content[inner_start..inner_end];
-                        let inner_children =
-                            self.parse_inline(inner_content, offset + inner_start)?;
-                        let span =
-                            Span::new((offset + pos) as u32, (offset + inner_end + 2) as u32);
-                        let delete = ox_content_ast::Delete { children: inner_children, span };
-                        children.push(Node::Delete(delete));
-                        pos = inner_end + 2;
-                    } else {
-                        let text = Text {
-                            value: &content[pos..pos + 2],
-                            span: Span::new((offset + pos) as u32, (offset + pos + 2) as u32),
-                        };
-                        children.push(Node::Text(text));
-                        pos += 2;
-                    }
-                }
-                b'*' | b'_' => {
-                    // Emphasis or Strong
-                    let marker = ch;
-                    let mut count = 1;
-                    while pos + count < content.len() && bytes[pos + count] == marker {
-                        count += 1;
-                    }
-
-                    // Simple logic: find next matching sequence of same length
-                    let inner_start = pos + count;
-                    let mut inner_end = inner_start;
-                    let mut found = false;
-
-                    while inner_end < content.len() {
-                        if bytes[inner_end] == marker {
-                            let mut end_count = 1;
-                            while inner_end + end_count < content.len()
-                                && bytes[inner_end + end_count] == marker
-                            {
-                                end_count += 1;
-                            }
-
-                            if end_count >= count {
-                                found = true;
-                                break;
-                            }
-                            inner_end += end_count;
-                        } else {
-                            inner_end += 1;
-                        }
-                    }
-
-                    if found {
-                        let inner_content = &content[inner_start..inner_end];
-                        // Recursively parse inner content
-                        let inner_children =
-                            self.parse_inline(inner_content, offset + inner_start)?;
-
-                        let span =
-                            Span::new((offset + pos) as u32, (offset + inner_end + count) as u32);
-
-                        if count >= 2 {
-                            // Strong
-                            let strong = ox_content_ast::Strong { children: inner_children, span };
-                            children.push(Node::Strong(strong));
-                            pos = inner_end + count;
-                        } else {
-                            // Emphasis
-                            let emphasis =
-                                ox_content_ast::Emphasis { children: inner_children, span };
-                            children.push(Node::Emphasis(emphasis));
-                            pos = inner_end + count;
-                        }
-                    } else {
-                        // Treat as text
-                        let text = Text {
-                            value: &content[pos..pos + count],
-                            span: Span::new((offset + pos) as u32, (offset + pos + count) as u32),
-                        };
-                        children.push(Node::Text(text));
-                        pos += count;
-                    }
-                }
-                b'`' => {
-                    // Inline code
-                    pos += 1;
-                    let code_start = pos;
-                    while pos < content.len() && bytes[pos] != b'`' {
-                        pos += 1;
-                    }
-                    if pos < content.len() {
-                        let code_content = &content[code_start..pos];
-                        let inline_code = ox_content_ast::InlineCode {
-                            value: code_content,
-                            span: Span::new(
-                                (offset + code_start - 1) as u32,
-                                (offset + pos + 1) as u32,
-                            ),
-                        };
-                        children.push(Node::InlineCode(inline_code));
-                        pos += 1;
-                    } else {
-                        // No closing backtick, treat as text
-                        let text = Text {
-                            value: &content[code_start - 1..],
-                            span: Span::new(
-                                (offset + code_start - 1) as u32,
-                                (offset + content.len()) as u32,
-                            ),
-                        };
-                        children.push(Node::Text(text));
-                    }
-                }
-                b'[' => {
-                    // Link: [text](url)
-                    let link_start = pos;
-                    pos += 1;
-                    let text_start = pos;
-
-                    // Find closing ]
-                    let mut bracket_depth = 1;
-                    while pos < content.len() && bracket_depth > 0 {
-                        match bytes[pos] {
-                            b'[' => bracket_depth += 1,
-                            b']' => bracket_depth -= 1,
-                            _ => {} // Ignore other characters
-                        }
-                        if bracket_depth > 0 {
-                            pos += 1;
-                        }
-                    }
-
-                    if pos < content.len()
-                        && bytes[pos] == b']'
-                        && pos + 1 < content.len()
-                        && bytes[pos + 1] == b'('
-                    {
-                        let link_text = &content[text_start..pos];
-                        pos += 2; // skip ](
-
-                        // Find closing )
-                        let url_start = pos;
-                        let mut paren_depth = 1;
-                        while pos < content.len() && paren_depth > 0 {
-                            match bytes[pos] {
-                                b'(' => paren_depth += 1,
-                                b')' => paren_depth -= 1,
-                                _ => {} // Ignore other characters
-                            }
-                            if paren_depth > 0 {
-                                pos += 1;
-                            }
-                        }
-
-                        if pos < content.len() && bytes[pos] == b')' {
-                            let url = &content[url_start..pos];
-                            pos += 1; // skip )
-
-                            // Parse link text as inline content
-                            let link_children =
-                                self.parse_inline(link_text, offset + text_start)?;
-
-                            let link = Link {
-                                url,
-                                title: None,
-                                children: link_children,
-                                span: Span::new(
-                                    (offset + link_start) as u32,
-                                    (offset + pos) as u32,
-                                ),
-                            };
-                            children.push(Node::Link(link));
-                        } else {
-                            // Invalid link, treat as text
-                            let text = Text {
-                                value: &content[link_start..pos],
-                                span: Span::new(
-                                    (offset + link_start) as u32,
-                                    (offset + pos) as u32,
-                                ),
-                            };
-                            children.push(Node::Text(text));
-                        }
-                    } else {
-                        // Not a link, just a [
-                        let text = Text {
-                            value: "[",
-                            span: Span::new(
-                                (offset + link_start) as u32,
-                                (offset + link_start + 1) as u32,
-                            ),
-                        };
-                        children.push(Node::Text(text));
-                        pos = link_start + 1;
-                    }
-                }
-                b'!' => {
-                    // Image: ![alt](url)
-                    if pos + 1 < content.len() && bytes[pos + 1] == b'[' {
-                        let image_start = pos;
-                        pos += 2; // skip ![
-                        let alt_start = pos;
-
-                        // Find closing ]
-                        let mut bracket_depth = 1;
-                        while pos < content.len() && bracket_depth > 0 {
-                            match bytes[pos] {
-                                b'[' => bracket_depth += 1,
-                                b']' => bracket_depth -= 1,
-                                _ => {}
-                            }
-                            if bracket_depth > 0 {
-                                pos += 1;
-                            }
-                        }
-
-                        if pos < content.len()
-                            && bytes[pos] == b']'
-                            && pos + 1 < content.len()
-                            && bytes[pos + 1] == b'('
-                        {
-                            let alt_text = &content[alt_start..pos];
-                            pos += 2; // skip ](
-
-                            // Find closing )
-                            let url_start = pos;
-                            let mut paren_depth = 1;
-                            while pos < content.len() && paren_depth > 0 {
-                                match bytes[pos] {
-                                    b'(' => paren_depth += 1,
-                                    b')' => paren_depth -= 1,
-                                    _ => {}
-                                }
-                                if paren_depth > 0 {
-                                    pos += 1;
-                                }
-                            }
-
-                            if pos < content.len() && bytes[pos] == b')' {
-                                let url = &content[url_start..pos];
-                                pos += 1; // skip )
-
-                                let image = Image {
-                                    url,
-                                    alt: alt_text,
-                                    title: None,
-                                    span: Span::new(
-                                        (offset + image_start) as u32,
-                                        (offset + pos) as u32,
-                                    ),
-                                };
-                                children.push(Node::Image(image));
-                            } else {
-                                // Invalid image, treat as text
-                                let text = Text {
-                                    value: &content[image_start..pos],
-                                    span: Span::new(
-                                        (offset + image_start) as u32,
-                                        (offset + pos) as u32,
-                                    ),
-                                };
-                                children.push(Node::Text(text));
-                            }
-                        } else {
-                            // Not an image, just ![
-                            let text = Text {
-                                value: "![",
-                                span: Span::new(
-                                    (offset + image_start) as u32,
-                                    (offset + image_start + 2) as u32,
-                                ),
-                            };
-                            children.push(Node::Text(text));
-                            pos = image_start + 2;
-                        }
-                    } else {
-                        // Just a !, treat as text
-                        let text = Text {
-                            value: "!",
-                            span: Span::new((offset + pos) as u32, (offset + pos + 1) as u32),
-                        };
-                        children.push(Node::Text(text));
-                        pos += 1;
-                    }
-                }
-                _ => {
-                    // Other special characters
-                    let text = Text {
-                        value: &content[pos..pos + 1],
-                        span: Span::new((offset + pos) as u32, (offset + pos + 1) as u32),
-                    };
-                    children.push(Node::Text(text));
-                    pos += 1;
-                }
-            }
-        }
-
-        Ok(children)
-    }
-
-    fn parse_inline_html(content: &'a str, pos: usize, offset: usize) -> Option<(Html<'a>, usize)> {
-        let bytes = content.as_bytes();
-
-        if content[pos..].starts_with("<!--") {
-            let end = content[pos + 4..].find("-->").map(|found| pos + 4 + found + 3)?;
-            return Some((
-                Html {
-                    value: &content[pos..end],
-                    span: Span::new((offset + pos) as u32, (offset + end) as u32),
-                },
-                end,
-            ));
-        }
-
-        let closing = bytes.get(pos + 1) == Some(&b'/');
-        let tag_start = if closing { pos + 2 } else { pos + 1 };
-        if !bytes.get(tag_start).is_some_and(u8::is_ascii_alphabetic) {
-            return None;
-        }
-
-        let mut tag_end = tag_start + 1;
-        while tag_end < bytes.len()
-            && (bytes[tag_end].is_ascii_alphanumeric() || bytes[tag_end] == b'-')
-        {
-            tag_end += 1;
-        }
-
-        let mut cursor = tag_end;
-        if closing {
-            while cursor < bytes.len() && matches!(bytes[cursor], b' ' | b'\t') {
-                cursor += 1;
-            }
-            if bytes.get(cursor) != Some(&b'>') {
-                return None;
-            }
-            let end = cursor + 1;
-            return Some((
-                Html {
-                    value: &content[pos..end],
-                    span: Span::new((offset + pos) as u32, (offset + end) as u32),
-                },
-                end,
-            ));
-        }
-
-        if !matches!(bytes.get(cursor), Some(b' ' | b'\t' | b'/' | b'>')) {
-            return None;
-        }
-
-        let mut quote = None;
-        while cursor < bytes.len() {
-            let byte = bytes[cursor];
-            if let Some(quote_byte) = quote {
-                if byte == quote_byte {
-                    quote = None;
-                }
-            } else if matches!(byte, b'"' | b'\'') {
-                quote = Some(byte);
-            } else if byte == b'>' {
-                let end = cursor + 1;
-                return Some((
-                    Html {
-                        value: &content[pos..end],
-                        span: Span::new((offset + pos) as u32, (offset + end) as u32),
-                    },
-                    end,
-                ));
-            } else if byte == b'\n' {
-                return None;
-            }
-            cursor += 1;
-        }
-
-        None
-    }
-}
-
-/// Lookup table: `INLINE_SPECIAL[b] == 1` iff the byte is one of the
-/// inline-special markers. Lookup tables vectorize and predict better
-/// than chained `matches!` checks.
-static INLINE_SPECIAL: [u8; 256] = {
-    let mut t = [0u8; 256];
-    t[b'*' as usize] = 1;
-    t[b'_' as usize] = 1;
-    t[b'`' as usize] = 1;
-    t[b'[' as usize] = 1;
-    t[b'!' as usize] = 1;
-    t[b'~' as usize] = 1;
-    t[b'\\' as usize] = 1;
-    t[b'<' as usize] = 1;
-    t
-};
-
-/// Returns the index of the next inline-special byte at or after `from`,
-/// or `bytes.len()` if none. Unrolled 8-byte chunks let LLVM vectorize the
-/// no-marker path — the common case in real-world Markdown is long runs
-/// of plain text between markers.
-#[inline]
-fn next_inline_special(bytes: &[u8], from: usize) -> usize {
-    let mut i = from;
-    let end = bytes.len();
-
-    while i + 8 <= end {
-        let chunk = &bytes[i..i + 8];
-        let mask = INLINE_SPECIAL[chunk[0] as usize]
-            | INLINE_SPECIAL[chunk[1] as usize]
-            | INLINE_SPECIAL[chunk[2] as usize]
-            | INLINE_SPECIAL[chunk[3] as usize]
-            | INLINE_SPECIAL[chunk[4] as usize]
-            | INLINE_SPECIAL[chunk[5] as usize]
-            | INLINE_SPECIAL[chunk[6] as usize]
-            | INLINE_SPECIAL[chunk[7] as usize];
-        if mask != 0 {
-            break;
-        }
-        i += 8;
-    }
-    while i < end && INLINE_SPECIAL[bytes[i] as usize] == 0 {
-        i += 1;
-    }
-    i
-}
-
-/// Cheap byte-level test mirroring `try_parse_heading_at` without bouncing
-/// through `chars()`/`peekable`. ATX heading prefix: `#{1..=6}` followed
-/// by space/tab/newline or end-of-input.
-fn is_atx_heading_prefix(bytes: &[u8]) -> bool {
-    let mut hashes = 0;
-    while hashes < bytes.len() && bytes[hashes] == b'#' {
-        hashes += 1;
-        if hashes > 6 {
-            return false;
-        }
-    }
-    if hashes == 0 {
-        return false;
-    }
-    matches!(bytes.get(hashes), None | Some(b' ' | b'\t' | b'\n'))
-}
-
-/// Case-insensitive search for `</tag` in `haystack`. Equivalent to
-/// `haystack.to_ascii_lowercase().contains(&format!("</{tag}"))` but
-/// allocates nothing. Used in the HTML block hot loop (type 1 closing
-/// detection) which iterates once per line inside `<pre>` / `<script>` /
-/// `<style>` / `<textarea>` blocks — historically the dominant allocator
-/// in this parser on documents with many such blocks.
-fn ascii_contains_closing_tag(haystack: &str, tag: &[u8]) -> bool {
-    let bytes = haystack.as_bytes();
-    if bytes.len() < tag.len() + 2 {
-        return false;
-    }
-    let limit = bytes.len() - (tag.len() + 1);
-    let mut i = 0;
-    while i <= limit {
-        if bytes[i] == b'<' && bytes[i + 1] == b'/' {
-            let candidate = &bytes[i + 2..i + 2 + tag.len()];
-            if candidate.eq_ignore_ascii_case(tag) {
-                return true;
-            }
-        }
-        i += 1;
-    }
-    false
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn ascii_contains_closing_tag_matches_case_insensitively() {
-        assert!(ascii_contains_closing_tag("end </SCRIPT> tail", b"script"));
-        assert!(ascii_contains_closing_tag("</style ", b"style"));
-        assert!(!ascii_contains_closing_tag("<scriptsource>", b"script"));
-        assert!(!ascii_contains_closing_tag("", b"pre"));
-        assert!(!ascii_contains_closing_tag("</pr", b"pre"));
-    }
-
-    #[test]
-    fn test_parse_image() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "![Alt text](/path/to/image.png)").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::Paragraph(p) => {
-                assert_eq!(p.children.len(), 1);
-                match &p.children[0] {
-                    Node::Image(img) => {
-                        assert_eq!(img.alt, "Alt text");
-                        assert_eq!(img.url, "/path/to/image.png");
-                    }
-                    _ => panic!("expected image, got {:?}", &p.children[0]),
-                }
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn test_parse_heading() {
-        let allocator = Allocator::new();
-        // Use "# " with trailing space - our parser requires space/tab/newline after #
-        let doc = Parser::new(&allocator, "# Hello\n").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::Heading(h) => {
-                assert_eq!(h.depth, 1);
-            }
-            _ => panic!("expected heading"),
-        }
-    }
-
-    #[test]
-    fn indented_heading_like_text_does_not_loop() {
-        // Regression: `line_starts_block` once tested ATX headings against
-        // the trimmed bytes while `parse_block` tested at the un-trimmed
-        // position, so " # heading" caused `parse_paragraph` to break
-        // immediately ("looks like a heading") and `parse_block` to return
-        // `Ok(None)` without advancing — spinning the outer loop forever.
-        // The fix is for `line_starts_block` to defer to
-        // `self.try_parse_heading_at()` so both checks see the same offset.
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, " # heading\n").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        assert!(
-            matches!(&doc.children[0], Node::Paragraph(_)),
-            "expected leading-indented `#` to parse as paragraph text, got {:?}",
-            &doc.children[0]
-        );
-    }
-
-    #[test]
-    fn test_parse_paragraph() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "Hello world").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        assert!(matches!(&doc.children[0], Node::Paragraph(_)));
-    }
-
-    #[test]
-    fn test_parse_thematic_break() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "---").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        assert!(matches!(&doc.children[0], Node::ThematicBreak(_)));
-    }
-
-    #[test]
-    fn test_parse_fenced_code() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(
-            &allocator,
-            "```rust\nfn main() {}
-```",
-        )
-        .parse()
-        .unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::CodeBlock(cb) => {
-                assert_eq!(cb.lang, Some("rust"));
-            }
-            _ => panic!("expected code block"),
-        }
-    }
-
-    #[test]
-    fn test_parse_inline_code() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "Use `code` here").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::Paragraph(p) => {
-                assert!(p.children.iter().any(|n| matches!(n, Node::InlineCode(_))));
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn test_parse_strikethrough() {
-        let allocator = Allocator::new();
-        let doc =
-            Parser::with_options(&allocator, "~~done~~", ParserOptions::gfm()).parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::Paragraph(p) => {
-                assert!(matches!(&p.children[0], Node::Delete(_)));
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn test_parse_hard_break() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "line 1\\\nline 2").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::Paragraph(p) => {
-                assert!(p.children.iter().any(|n| matches!(n, Node::Break(_))));
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn test_parse_table() {
-        let allocator = Allocator::new();
-        let table_md = "| Header 1 | Header 2 |\n|----------|----------|\n| Cell 1   | Cell 2   |";
-        let parser = Parser::with_options(&allocator, table_md, ParserOptions::gfm());
-        let doc = parser.parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::Table(t) => {
-                assert_eq!(t.children.len(), 2); // header + 1 body row
-            }
-            _ => panic!("expected table, got {:?}", &doc.children[0]),
-        }
-    }
-
-    #[test]
-    fn test_parse_unordered_list() {
-        let allocator = Allocator::new();
-        let list_md = "- Item 1\n- Item 2\n- Item 3";
-        let parser = Parser::new(&allocator, list_md);
-        let doc = parser.parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::List(list) => {
-                assert!(!list.ordered);
-                assert_eq!(list.children.len(), 3);
-            }
-            _ => panic!("expected list, got {:?}", &doc.children[0]),
-        }
-    }
-
-    #[test]
-    fn test_parse_ordered_list() {
-        let allocator = Allocator::new();
-        let list_md = "1. First\n2. Second\n3. Third";
-        let parser = Parser::new(&allocator, list_md);
-        let doc = parser.parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::List(list) => {
-                assert!(list.ordered);
-                assert_eq!(list.children.len(), 3);
-            }
-            _ => panic!("expected list, got {:?}", &doc.children[0]),
-        }
-    }
-
-    #[test]
-    fn test_parse_block_quote() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "> Hello world").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::BlockQuote(bq) => {
-                assert_eq!(bq.children.len(), 1);
-                assert!(matches!(&bq.children[0], Node::Paragraph(_)));
-            }
-            _ => panic!("expected block quote, got {:?}", &doc.children[0]),
-        }
-    }
-
-    #[test]
-    fn test_parse_block_quote_multiline() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "> line 1\n> line 2").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::BlockQuote(bq) => {
-                assert_eq!(bq.children.len(), 1);
-            }
-            _ => panic!("expected block quote, got {:?}", &doc.children[0]),
-        }
-    }
-
-    #[test]
-    fn test_parse_nested_block_quote() {
-        let allocator = Allocator::new();
-        let doc = Parser::new(&allocator, "> > nested").parse().unwrap();
-        assert_eq!(doc.children.len(), 1);
-        match &doc.children[0] {
-            Node::BlockQuote(bq) => {
-                assert_eq!(bq.children.len(), 1);
-                assert!(matches!(&bq.children[0], Node::BlockQuote(_)));
-            }
-            _ => panic!("expected block quote, got {:?}", &doc.children[0]),
-        }
     }
 }
